Move StorageNodeType
