+++ crates/storage-ops/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod pruning;
 pub mod rollback;
+pub mod types;
 
 pub(crate) mod macros;
 #[cfg(test)]
