+++ crates/storage-ops/src/pruning/mod.rs
@@ -4,16 +4,15 @@ use futures::future;
 use serde::{Deserialize, Serialize};
 use sov_db::schema::tables::{LastPrunedBlock, LastPrunedL2Height};
 use tracing::info;
-use types::StorageNodeType;
 
 use self::components::{prune_ledger, prune_native_db};
 use self::criteria::{Criteria, DistanceCriteria};
 pub use self::service::*;
+use crate::types::StorageNodeType;
 
 pub(crate) mod components;
 pub(crate) mod criteria;
 pub(crate) mod service;
-pub mod types;
 
 /// A configuration type to define the behaviour of the pruner.
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
