+++ crates/storage-ops/src/types.rs
@@ -1,4 +1,6 @@
-#[derive(Copy, Clone)]
+use derive_more::Display;
+
+#[derive(Copy, Clone, Display)]
 pub enum StorageNodeType {
     Sequencer,
     FullNode,
